+++ app/controller/ember_cli/ember_controller.rb
@@ -3,10 +3,5 @@ module EmberCli
     def index
       render layout: false
     end
-
-    def ember_app
-      params[:ember_app]
-    end
-    helper_method :ember_app
   end
 end