+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 master
 ------
 
+* `EmberCli::EmberController` now inherits from `ActionController::Base` instead
+  of `ApplicationController`. [#400]
+* Remove support for Ruby 2.1.x. [#400]
 * Don't route requests to `/rails/info` through the mounted Ember application.
 
 0.7.4
@@ -33,6 +36,7 @@ master
 * `EmberCli::Deploy::File` serves assets with Rails' `static_cache_control`
   value. [#403]
 
+[#400]: https://github.com/thoughtbot/ember-cli-rails/pull/400
 [#396]: https://github.com/thoughtbot/ember-cli-rails/pull/396
 [#403]: https://github.com/thoughtbot/ember-cli-rails/pull/403